Fix missing Cancel button on iPad confirmation dialogs
