+++ Packages/Components/Sources/Extensions/SwiftUIView.swift
@@ -68,55 +68,6 @@ public extension View {
     }
 }
 
-// MARK: - Confirmation Dialog
-
-public extension View {
-    func confirmationDialog<S, A, T, M>(
-        _ title: S,
-        presenting data: Binding<T?>,
-        sensoryFeedback: SensoryFeedback? = nil,
-        @ViewBuilder actions: (T) -> A,
-        @ViewBuilder message: () -> M = { EmptyView() }
-    )
-    -> some View where S: StringProtocol, A: View, M: View {
-        let isPresented: Binding<Bool> = Binding(
-            get: { data.wrappedValue != nil },
-            set: { newValue in
-                guard !newValue else { return }
-                data.wrappedValue = nil
-            }
-        )
-        // confiramtion dialog works good only for iPhone, for different devices use an alert
-        let iPhone = UIDevice.current.userInterfaceIdiom == .phone
-
-        return ifElse(iPhone) {
-            $0.confirmationDialog(
-                    title,
-                    isPresented: isPresented,
-                    titleVisibility: .visible,
-                    presenting: data.wrappedValue,
-                    actions: actions,
-                    message: { _ in
-                        message()
-                    }
-                )
-        } elseContent: {
-            $0.alert(
-                    title,
-                    isPresented: isPresented,
-                    presenting: data.wrappedValue,
-                    actions: actions,
-                    message: { _ in
-                        message()
-                    }
-                )
-        }
-        .ifLet(sensoryFeedback) { view, value in
-            view.sensoryFeedback(value, trigger: isPresented.wrappedValue) { $1 }
-        }
-    }
-}
-
 // MARK: - Sheet
 
 public extension View {
